Rewrite logic to use a timeout
Now using a state machine to make it a bit easier to reason about.

When Space is held down it goes into the DECIDE state.
During DECIDE, any other key-down events are buffered.
A key-up of a buffered key (ie press and release during DECIDE) results
in a transition to SHIFT, as does holding down Space for 200ms.
Releasing Space without either of these happening causes it to send
space down and up, then replay any buffered key-down events. The latter
is most important for comfort and not reordering keys...